% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bg.pts.R
\docType{data}
\name{bg.pts}
\alias{bg.pts}
\title{Block group internal points from Census Bureau for 2020 geographies}
\format{
A data.frame of block groups, with these variables: \cr\cr
\preformatted{
'data.frame':	242335 obs. of  5 variables:
 $ FIPS      : chr  "010010201001" "010010201002" "010010202001" "010010202002" ...
 $ lat       : num  32.5 32.5 32.5 32.5 32.5 ...
 $ lon       : num  -86.5 -86.5 -86.5 -86.5 -86.5 ...
 $ bgid      : int  1 2 3 4 5 6 7 8 9 10 ...
 $ blockcount: int  26 30 14 27 21 21 16 29 29 18 ...
 - attr(*, "census_version")= num 2020
 - attr(*, "acs_version")= chr "2016-2020"
 - attr(*, "acs_releasedate")= chr "3/17/2022"
 - attr(*, "ejscreen_version")= chr "2.1"
 - attr(*, "ejscreen_releasedate")= chr "October 2022"
 - attr(*, "ejscreen_pkg_data")= chr "bg22"
}
}
\source{
Derived from Census Bureau, obtained 10/2022. 
  bg.pts <- as.data.frame(copy(EJAM::bgpts)) # see EJAM package for details 
  names(bg.pts) <- c('bgid', 'lat','lon','FIPS','blockcount')
  bg.pts <- bg.pts[ , c('FIPS', 'lat', 'lon', 'bgid', 'blockcount')]
  bg.pts <- metadata_add(bg.pts)
  usethis::use_data(bg.pts)
}
\description{
The population weighted mean of block internal points, 
  within each block group.
}
\details{
This data set provides a data.frame with 242,335 block groups, with 5 variables.
  It has PR but lacks "AS" "GU" "MP" "VI" 
  The population weighted mean not spatial mean was used
  to summarize the lat lon values of all Census 2020 blocks in a given block group.
}
\seealso{
\pkg{EJAM::proxistat2}  \pkg{EJAM::bgpts} and  \pkg{acs}  and for older data \pkg{UScensus2010blocks}
}
\keyword{datasets}
