% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockpoints_area_pop.R
\docType{data}
\name{blockpoints_area_pop}
\alias{blockpoints_area_pop}
\title{Block points from Census Bureau for 2020 geographies}
\source{
Derived from Census Bureau, obtained 10/2022.
}
\description{
data.table columns are: blockfips, lat, lon, pop, area, blockid
}
\details{
Area is in square meters. 
  This data set provides a data.table with 8,132,968 blocks  
  It has PR but lacks "AS" "GU" "MP" "VI"
}
