% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bufferapi.R
\name{bufferapi}
\alias{bufferapi}
\title{Use EJSCREEN API to get buffer results (conditions near each site)}
\usage{
bufferapi(lon, lat, radius = 5)
}
\arguments{
\item{lon}{vector of longitudes in decimal degrees}

\item{lat}{vector of latitudes in decimal degrees}

\item{radius}{single number, for all buffer circles (distance from site) (meters?)}
}
\value{
list of data.frames, one per buffer (one per site defined by a lat lon pair)
}
\description{
Use EJSCREEN API to get buffer results (conditions near each site)
}
