% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.dbfs.R
\name{compile.dbfs}
\alias{compile.dbfs}
\title{Read a series of dbf files and join compile them as a single data.frame}
\usage{
compile.dbfs(filepaths)
}
\arguments{
\item{filepaths}{Required charater vector with full paths including file names of dbf files to be read.}
}
\value{
Returns a data.frame with as many columns as each dbf file, and as many rows as there are in all the dbf files read in.
}
\description{
Can help with a series of downloaded Census data files, such as one file per State. 
  Reads each using \code{\link[foreign]{read.dbf}} and combines them all as a data.frame.
  They must all have the same columns, and each file just provides additional rows of data.
}
\seealso{
\code{\link[foreign]{read.dbf}}
}
