% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/deg2rad.R
\name{deg2rad}
\alias{deg2rad}
\title{Convert degrees to radians}
\usage{
deg2rad(deg)
}
\arguments{
\item{deg}{Degrees as decimal degrees, required numeric vector}
}
\value{
Radians, as numeric vector same length as input
}
\description{
From \url{http://www.r-bloggers.com/great-circle-distance-calculations-in-r/}
}

