% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ejscreenapi.R
\name{ejscreenapi}
\alias{ejscreenapi}
\title{Use EJSCREEN API to get stats on each circular buffer}
\usage{
ejscreenapi(lon, lat, radius = 5)
}
\arguments{
\item{lon}{Longitude numeric vector}

\item{lat}{Latitude numeric vector}

\item{radius}{radius of circular buffer - Probably in miles but should confirm}
}
\description{
Requests a standard EJSCREEN report for each of one or more circular buffers.
Specify a radius and vector of latitude longitude points,
and get for a buffer the population weighted mean value of each raw indicator
like percent low-income, and total population count, and percentiles for those
raw indicator scores, all from EJSCREEN, as in an EJSCREEN standard report.
}
\examples{
 
 \dontrun{
 # Specify size of buffer circle and pick random points as example data
 myradius <- 3
 n <- 10
 pts <- proxistat::testpoints_block2010(n)
 
 benchmark.start <- Sys.time()
  out <- proxistat::ejscreenapi(pts$lon, lat=pts$lat, radius = myradius)
 benchmark.end <- Sys.time()
  
 # Average speed
 perhour <- proxistat::speedsummary(benchmark.start, benchmark.end, NROW(pts))
 cat('\n', perhour, 'buffers per hour \n')
 out[,c(1:18,162:173)]
 t(out[1:2,]) 
 }
}
