% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gcd.R
\name{gcd}
\alias{gcd}
\title{Distance between two points by Haversine, Spherical Law of Cosines, or Vincenty inverse formula}
\usage{
gcd(long1, lat1, long2, lat2, algorithm = "haversine")
}
\arguments{
\item{long1}{longitude in decimal degrees}

\item{lat1}{latitude in decimal degrees}

\item{long2}{longitude in decimal degrees}

\item{lat2}{latitude in decimal degrees}

\item{algorithm}{Character string "hf" for Haversine by default, or "slc" for Spherical Law of Cosines, or "vif" for Vincenty inverse formula}
}
\value{
Distance in kilometers
}
\description{
Calculates the geodesic distance between two points specified by degrees (DD) latitude/longitude using
Haversine formula (hf), Spherical Law of Cosines (slc) and Vincenty inverse formula for ellipsoids (vif)
Taken from \url{http://www.r-bloggers.com/great-circle-distance-calculations-in-r/}
Note these are not the most accurate method for long distances (e.g., >1000 km) nor the fastest.
}

