% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gcd.slc.R
\name{gcd.slc}
\alias{gcd.slc}
\title{Distance based on spherical law of cosines}
\usage{
gcd.slc(long1, lat1, long2, lat2)
}
\arguments{
\item{long1}{longitude(s) in radians}

\item{lat1}{latitude(s) in radians}

\item{long2}{longitude(s) in radians}

\item{lat2}{latitude(s) in radians}
}
\value{
Distance in km
}
\description{
Calculates the geodesic distance between two points specified by radian latitude/longitude
using the Spherical Law of Cosines (slc).
Taken from \url{http://www.r-bloggers.com/great-circle-distance-calculations-in-r/}
}

