% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get.bg.latlons.R
\name{get.bg.latlons}
\alias{get.bg.latlons}
\title{Get block group internal points directly from Census shapefiles}
\usage{
get.bg.latlons(myyear = 2014, mytempdir, mystatenums, overwrite = FALSE,
  silent = FALSE)
}
\arguments{
\item{myyear}{Year such as 2014 (default)}

\item{mytempdir}{Optional. Default is TIGERTEMP created inside the current working directory. Character string of path to where temporary directory is or will be, for downloaded zip and dbf files.}

\item{mystatenums}{Optional, default is all including PR/VI/DC, etc. Vector of strings of 2-character FIPS codes for states to get data for.}

\item{overwrite}{Optional, FALSE by default (unlike in unzip), which means
if zip file exists locally do not download and if contents exist do not unzip.
BUT note this does not yet check to see if contents exist and zip does not, in which case could avoid downloading.}

\item{silent}{Optional logical FALSE by default. If TRUE, print more results.}
}
\description{
Download, unzip, read, and assemble lat lon of internal points for US Census block groups.
}
\examples{
#get.bg.latlons(mystatenums=c('09','10'))
}

