% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get.nearest.R
\name{get.nearest}
\alias{get.nearest}
\title{Find distance from each point in a set to the nearest of a second set of points (by lat/lon).}
\usage{
get.nearest(frompoints, topoints, return.units = "miles",
  return.rownums = TRUE, return.latlons = FALSE)
}
\arguments{
\item{frompoints}{A matrix or data.frame with two cols, 'lat' and 'lon' (or only 2 cols that are lat and lon in that order) with datum=WGS84 assumed.}

\item{topoints}{A matrix or data.frame with two cols, 'lat' and 'lon' (or only 2 cols that are lat and lon in that order) with datum=WGS84 assumed.}

\item{return.units}{A string that is 'miles' by default, or 'km' for kilometers, specifying units for distances returned.}

\item{return.rownums}{Logical value, TRUE by default. If TRUE, value returned also includes these 2 columns:
a col named fromrow of index numbers starting at 1 specifying the frompoint and a similar col named n specifying the row of the nearest topoint.}

\item{return.latlons}{Logical value, FALSE by default. If TRUE, value returned also includes four extra columns,
showing fromlat, fromlon, tolat, tolon.}
}
\value{
By default, returns a matrix of numbers, with columns fromrow, n indexing which is nearest of topoints, and d (distance).
  Distance returned is in miles by default, but with option to set return.units='km' to get kilometers.
  See parameters for details on other formats that may be returned if specified.
}
\description{
\code{get.nearest} returns the distance from each point in a set to the nearest of a second set of points (by lat/lon).
}
\details{
This function returns a vector of distances,
which are the distances from one set of points to the nearest single member (if any) of another set of points.
Points are specified using latitude and longitude in decimal degrees.
Relies on the sp package for the \code{\link{spDists}} and \code{\link{SpatialPoints}} functions.
A future version may use get.distances.all() but for performance only use it for distance pairs (pairs of points) that have been initially
quickly filtered using lat/lon to be not too far, in an attempt to go much faster in an initial pass.
}
\examples{
test.from <- structure(list(fromlat = c(38.9567309094, 38.9507043428),
 fromlon = c(-77.0896572305, -77.0896199948)), .Names = c("lat", "lon"),
 row.names = c("6054762", "6054764"), class = "data.frame")
test.to <- structure(list(tolat = c(38.9575019287, 38.9507043428, 38.9514152435),
 tolon = c(-77.0892818598, -77.0896199948, -77.0972395245)), .Names = c("lat", "lon"),
 class = "data.frame", row.names = c("6054762", "6054763", "6054764"))
get.nearest(test.from, test.to)
}
\seealso{
\code{\link{get.distances}} for get.distances() which gets distances between all points (within an optional search radius),
  \code{\link{get.distances.all}} for get.distances.all() which allows you to get distances between all points,
  \code{\link{get.distances.prepaired}} for finding distances when data are already formatted as pairs of points, and
  \code{\link{proxistat}} which calculates a proximity score for each spatial unit based on distances to nearby points.
}
\concept{
proximity
}

