% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlon.colnames.check.R
\name{latlon.colnames.check}
\alias{latlon.colnames.check}
\title{Utility function to check for valid lat/lon columns}
\usage{
latlon.colnames.check(mypoints)
}
\arguments{
\item{mypoints}{A matrix or data.frame}
}
\value{
Returns a vector of colnames such as c('lat', 'lon') or stops if problem found
}
\description{
Used by functions like \code{\link{get.distances}} to check input parameters frompoints and topoints
}
\examples{
#
}
\seealso{
\code{\link{get.distances.all}} which allows you to get distances between all points,
  \code{\link{get.distances.prepaired}} for finding distances when data are already formatted as pairs of points,
  \code{\link{get.nearest}} which finds the distance to the single nearest point 
  within a specified search radius instead of all topoints, and 
  \code{\link{proxistat}} which calculates a proximity score for each spatial unit based on distances to nearby points.

\code{\link{get.distances}} which allows you to specify a search radius and 
  get distances only within that radius which can be faster,
  \code{\link{get.distances.prepaired}} for finding distances when data are already formatted as pairs of points,
  \code{\link{get.nearest}} which finds the distance to the single nearest point 
  within a specified search radius instead of all topoints, and 
  \code{\link{proxistat}} which calculates a proximity score for each spatial unit
}
\concept{
proximity
}
