% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlon_is.valid.R
\name{latlon_is.valid}
\alias{latlon_is.valid}
\title{Validate latitudes and longitudes}
\usage{
latlon_is.valid(lat, lon)
}
\arguments{
\item{lat}{vector of latitudes in decimal degrees}

\item{lon}{numeric vector of longitudes in decimal degrees, same length}
}
\value{
logical vector, one element per lat lon pair (location)
}
\description{
Check each latitude and longitude value to see if they are 
  NA or outside expected numeric ranges
}
\seealso{
latlon_df_clean() latlon_infer() latlon_is.valid() latlon_as.numeric()
}
