% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/linesegments.R
\name{linesegments}
\alias{linesegments}
\title{Add line segments connecting pairs of points, to a plot}
\usage{
linesegments(xfrom, yfrom, xto, yto, ...)
}
\arguments{
\item{xfrom}{required numeric vector of x values for starting points}

\item{yfrom}{required numeric vector of y values for starting points}

\item{xto}{required numeric vector of x values for ending points}

\item{yto}{required numeric vector of y values for ending points}

\item{...}{optional additional parameters to pass to \code{\link{lines}}}
}
\value{
Draws lines(), one line segment from each starting point to its corresponding ending point.
}
\description{
Accepts vectors of x and y values of pairs of points (e.g., longitude and latitude),
  reformats them and adds line segments to an existing plot. Each line segment drawn
  connects one point its paired point. The two sets of points have to be equal in length, set up as pairs.
}
\details{
This function also silently returns a matrix of two columns.
  Each column is a vector that has elements sequenced in groups of three --
  the from point, then the to point, then NA to signify a break in line drawing. See \code{\link{lines}}
}
\examples{
t10  <- testpoints(10,  minlat = 25, maxlat = 45, minlon = -100, maxlon = -60)
 t100 <- testpoints(100, minlat = 25, maxlat = 45, minlon = -100, maxlon = -60)
 nears=as.data.frame( get.nearest(t10, t100, return.latlons=TRUE) )
 plot(t10)
 plot(t100, pch='.')
 linesegments(nears$fromlon, nears$fromlat, nears$tolon, nears$tolat)
}
\seealso{
\code{\link{testpoints}} and \code{\link{get.nearest}}
}

