% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/meters.per.degree.lat.R
\name{meters.per.degree.lat}
\alias{meters.per.degree.lat}
\title{Convert degrees latitude to meters North-South.}
\usage{
meters.per.degree.lat(theta)
}
\arguments{
\item{theta}{The decimal degrees of latitude of the Northern Hemisphere location(s) of interest, as number or vector of numbers.}
}
\value{
Returns meters traveled N-S, as a number or vector of numbers the same length as the input.
}
\description{
\code{meters.per.degree.lat} returns meters traveled North-South per decimal degrees latitude, given latitude (Northern Hemisphere).
}
\details{
This function returns the meters traveled North-South per decimal degree latitude, at a given latitude (Northern Hemisphere).
This is an approximation and is less accurate further from the given latitude.
Based on
\url{http://en.wikipedia.org/wiki/Latitude#Length_of_a_degree_of_latitude} and
\url{http://en.wikipedia.org/wiki/Longitude#Length_of_a_degree_of_longitude}
\cr
Input theta is latitude on WGS84.
\cr
Also see \url{http://en.wikipedia.org/wiki/Longitude} and \url{http://en.wikipedia.org/wiki/Decimal_degrees}
}
\examples{
meters.per.degree.lat(32)
meters.per.degree.lat(c(0,45,72))
}
\seealso{
\code{\link{meters.per.degree.lon}} for a similar function but for travel East-West, with more detailed explanation/help,
  and \code{\link{get.distances.all}} and \code{\link{get.distances}} for distances between points, and related functions.
}
\concept{
proximity
}

