% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/near_eachother.R
\name{near_eachother}
\alias{near_eachother}
\title{which points are near any of the others in list?}
\usage{
near_eachother(lon, lat, distance, or_tied = FALSE)
}
\arguments{
\item{lon}{longitude}

\item{lat}{latitude}

\item{distance}{distance between points in miles to check}

\item{or_tied}{if TRUE, checks if less than or equal to distance, otherwise if less than}
}
\description{
which points are near any of the others in list?
}
