% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pointmap.R
\name{pointmap}
\alias{pointmap}
\title{Read a series of dbf files and join compile them as a single data.frame}
\usage{
pointmap(bin, lat, lon, vartext = "x", areatext = "area",
  coloring = colors[1:length(unique(bin))], asp = c(1, 1), pch = 1, ...)
}
\arguments{
\item{bin}{}

\item{lat}{}

\item{lon}{}

\item{vartext}{default is 'x', text to use in describing the field mapped}

\item{areatext}{default is 'area', text to use in describing the area mapped}

\item{coloring}{default is some basic colors, vector specifying color for each bin}

\item{asp}{default is c(1,1), aspect ratio of graphic}

\item{pch}{default is 1, defines type of marker for each point}

\item{...}{other params to send to plot}
}
\value{
Draws a map
}
\description{
Can help with a series of downloaded Census data files, such as one file per State.
  Reads each using \code{\link[foreign]{read.dbf}} and combines them all as a data.frame.
  They must all have the same columns, and each file just provides additional rows of data.
}
\seealso{
\code{\link{countypointmap}}
}

