\name{proxistat-package}
\alias{proxistat-package}
\alias{proxistat}
\docType{package}
\title{
\packageTitle{proxistat}
}
\description{
\packageDescription{proxistat}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{proxistat}
\packageIndices{proxistat}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{proxistat}

Maintainer: \packageMaintainer{proxistat}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
