% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxistat.assemble.chunks.R
\name{proxistat.assemble.chunks}
\alias{proxistat.assemble.chunks}
\title{read files storing proxistat result chunks and assemble into one result}
\usage{
proxistat.assemble.chunks(files, folder = getwd())
}
\arguments{
\item{files}{Required character vector of names of .RData files to read and combine}

\item{folder}{Optional character element specifying directory where files are stored, defaults to getwd()}
}
\value{
Matrix that contains combined results found in all the files
}
\description{
read files storing proxistat result chunks and assemble into one result
}
\examples{
\dontrun{
 fnames=proxistat.chunked(testpoints(10), testpoints(5), fromchunksize = 4, assemble=FALSE, 
   folder=file.path(getwd(), 'temp'))
 output=proxistat.assemble.chunks(files=fnames, folder=file.path(getwd(), 'temp'))
  }
}
