% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speedsummary.R
\name{speedsummary}
\alias{speedsummary}
\title{Print summary of speed in buffers per hour, given start and end times of batch}
\usage{
speedsummary(start, end, n)
}
\arguments{
\item{start}{time when finished from Sys.time()}

\item{end}{time when finished from Sys.time()}

\item{n}{how many buffers were done?}
}
\value{
Prints to console and returns rate of buffers per hour
}
\description{
Print summary of speed in buffers per hour, given start and end times of batch
}
\examples{
\dontrun{
 pts <- proxistat::testpoints_block2010(10)
 benchmark.start <- Sys.time()
 outlist <- batch.summarizer::ejscreenapi(pts$lon, lat=pts$lat, radius = 5)
 benchmark.end <- Sys.time()
 proxistat::speedsummary(benchmark.start, benchmark.end, NROW(pts))
}

}
