% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testpoints_bg20.R
\name{testpoints_bg20}
\alias{testpoints_bg20}
\title{Get lat lon for random US census block groups (where avg person lives or other weighting)}
\usage{
testpoints_bg20(size, popwtd = TRUE)
}
\arguments{
\item{size}{number of unique test points needed (no duplicates will be returned)}

\item{popwtd}{default is TRUE, and pop weighted points represent block group where the avg US resident lives.
If FALSE, unweighted, so each block group has equal chance of being selected.}
}
\value{
data.frame with columns lon and lat, in decimal degrees
}
\description{
Get lat lon for random US census block groups (where avg person lives or other weighting)
}
\details{
Note this is not optimized for speed. can be slow if pop wtd. 
 May recode this to allow wts to be pop, area, or none (no weight).
 Area weighted would mean a random location means each square meter is equally likely to be picked, 
 like throwing a dart at a map except it is rounded to the nearest internal point. 
 Since block groups generally almost all tend to have somewhat similar pop counts,
 popwtd and bg fips wtd are not all that different. But square meters varies a lot, 
 so area weighted is very different than pop or fips weighted.
}
\examples{
#library(jsonlite)
#library(httr)
#library(tidyverse) # need magrittr
#library(data.table)
#library(sf)
# pts <- proxistat::testpoints_bg20(10)
# benchmark.start <- Sys.time()
# outlist <- proxistat::bufferapi(pts$lon, lat=pts$lat, radius = 5)
# benchmark.end <- Sys.time()
# proxistat::speedsummary(benchmark.start, benchmark.end, NROW(pts))
 # VISUALIZE variability in speed
 # hist(as.numeric(unlist(lapply(outlist, FUN =function(x) x$timeSeconds)), na.rm = T),100, 
 #   main = 'Histogram of seconds per buffer', xlab='Seconds elapsed for a buffer query of API')
 
}
