% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testpoints_block2010.R
\name{testpoints_block2010}
\alias{testpoints_block2010}
\title{Get lat lon for random US census blocks (where avg person lives or other weighting)}
\usage{
testpoints_block2010(size, popwtd = TRUE)
}
\arguments{
\item{size}{number of unique test points needed (no duplicates will be returned)}

\item{popwtd}{default is TRUE, and pop weighted block points represent where the avg US resident lives.
If FALSE, unweighted, so each block has equal chance of being selected.}
}
\value{
data.frame with columns lon and lat, in decimal degrees
}
\description{
Note the 2010 version of this is not optimized for speed- it can take 30 seconds to pick just 1,000 points if pop wtd. 
 May recode this to allow wts to be pop (using blocks.pop), or area (using blocks.area), or none (no weight).
}
\details{
This is probably obsolete at this point.
}
