% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testpoints_block2020.R
\name{testpoints_block2020}
\alias{testpoints_block2020}
\title{Get lat lon pop and area for random US census blocks (where avg person lives or other weighting)}
\usage{
testpoints_block2020(size, weighting = "area")
}
\arguments{
\item{size}{number of unique test points needed (no duplicates will be returned)}

\item{weighting}{is "people" for average resident; or "block" for average census block,
or "area" for average square meter 
If FALSE, unweighted, so each block has equal chance of being selected.}
}
\value{
data.frame with columns lon and lat, in decimal degrees, and area and pop count
}
\description{
Get lat lon pop and area for random US census blocks (where avg person lives or other weighting)
}
