% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bufferapi.R
\name{bufferapi}
\alias{bufferapi}
\title{Use the EJSCREEN API to get standard report for each of one or more circular buffers}
\usage{
bufferapi(lon, lat, radius = 5)
}
\arguments{
\item{lon}{Longitude numeric vector}

\item{lat}{Latitude numeric vector}

\item{radius}{radius of circular buffer - I think in miles but should confirm}
}
\description{
Use the EJSCREEN API to get standard report for each of one or more circular buffers
}
\examples{
 
  \dontrun{
  sample.count <- 10
  pts <- proxistat::testpoints_block2010(sample.count)
  outlist <- proxistat::bufferapi(pts$lon, lat=pts$lat, radius = 3)
  summary(outlist)
  class(outlist[[1]])
  names(outlist[[1]])
  x <- do.call(rbind, outlist)
  x[,c(1:18,162:173)]
  t(x[1:2,]) 
  }
}
