% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countypointmap.R
\name{countypointmap}
\alias{countypointmap}
\title{Simple map of color-coded block group centroids in county}
\usage{
countypointmap(query, vartext, varname, breaks = 5, coloring, asp = c(1, 1),
  pch = 1, ......)
}
\arguments{
\item{query}{See \code{\link[ejanalysis]{get.county.info}}}
}
\value{
Just draws a map
}
\description{
very simple mapping
}
\seealso{
\code{\link{pointmap}}
}

