% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gcd.hf.R
\name{gcd.hf}
\alias{gcd.hf}
\title{Distance based on Haversine formula}
\usage{
gcd.hf(long1, lat1, long2, lat2)
}
\arguments{
\item{long1}{longitude(s) in radians}

\item{lat1}{latitude(s) in radians}

\item{long2}{longitude(s) in radians}

\item{lat2}{latitude(s) in radians}
}
\value{
Distance in kilometers
}
\description{
Calculates the geodesic distance between two points (or multiple pairs of points) specified by radian latitude/longitude
using the Haversine formula (hf)
Taken from \url{http://www.r-bloggers.com/great-circle-distance-calculations-in-r/}
but use pmin instead of min to vectorize it.
}

