% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gcd.R
\name{gcd}
\alias{gcd}
\title{Distance between two points by Haversine, Spherical Law of Cosines, or Vincenty inverse formula}
\usage{
gcd(frompoints, topoints, dfunc = "hf", units = "km")
}
\arguments{
\item{frompoints}{Required matrix or data.frame of 2 columns, named lat and lon, with latitude and longitude(s) in degrees, one row per point.}

\item{topoints}{Required matrix or data.frame of 2 columns, named lat and lon, with latitude and longitude(s) in degrees, one row per point.}

\item{dfunc}{Character string "hf" for Haversine by default, or "slc" for Spherical Law of Cosines
(and may add "vif" for Vincenty inverse formula but that is not implemented here).
For "sp" algorithm (\pkg{sp}), see \code{\link{get.distances}}}

\item{units}{Optional character variable specifying 'km' or 'miles', 'km' by default.}
}
\value{
Distance in kilometers by default, or in miles if units='miles'
}
\description{
Calculates the geodesic distance between two points (or multiple pairs of points) specified by degrees (DD) latitude/longitude using
Haversine formula (hf), Spherical Law of Cosines (slc) and Vincenty inverse formula for ellipsoids (vif) \cr
Taken from \url{http://www.r-bloggers.com/great-circle-distance-calculations-in-r/} \cr
Note these are not the most accurate method for long distances (e.g., >1000 km) nor the fastest.
}
\details{
*** frompoints and topoints must have same number of rows, defining all pairs of points.
  Alternatively can define a single point while topoints defines a series of points, or vice versa.
  Taken from \url{http://www.r-bloggers.com/great-circle-distance-calculations-in-r/}
  but use \code{\link{pmin}} instead of \code{\link{min}} to vectorize it to handle at least pairs,
  and parameters changed to keep lat/lon as a matrix or data.frame rather than 2 separate vectors.
}
\seealso{
\code{\link{convert}}, \code{\link{gcd}}, \code{\link{get.distances}}, \code{\link{get.distances.all}}
}

