% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointmap.R
\name{pointmap}
\alias{pointmap}
\title{Read a series of dbf files and join compile them as a single data.frame}
\usage{
pointmap(bin, lat, lon, vartext = "x", areatext = "area",
  coloring = rainbow(length(unique(bin))), asp = c(1, 1), pch = ".", ...)
}
\arguments{
\item{bin}{Indicates color of each point. A vector of integers used to index the coloring vector. That would ideally include just 1:n where n is the n number of unique values and the index to the coloring vector, so each unique value 1:5, for example, is assigned a map color that is coloring[bin]}

\item{lat}{vector of latitudes just interpreted as y values to plot}

\item{lon}{vector of longitudes just interpreted as x values to plot}

\item{vartext}{default is 'x', text to use in describing the field mapped}

\item{areatext}{default is 'area', text to use in describing the area mapped}

\item{coloring}{default is some basic colors from rainbow(). A vector specifying color for each bin (should be same length as number of unique bin values)}

\item{asp}{default is c(1,1), aspect ratio of graphic (not right for AK, e.g.)}

\item{pch}{default is '.', see \code{\link{par}} -- defines type of marker for each point}

\item{...}{other params to send to plot}
}
\value{
Draws a map
}
\description{
Can help with a series of downloaded Census data files, such as one file per State. 
  Reads each using \code{\link[foreign]{read.dbf}} and combines them all as a data.frame.
  They must all have the same columns, and each file just provides additional rows of data.
}
\examples{
 
  myfips <- bg.pts$FIPS[substr(bg.pts$FIPS,1,2)=='06'] # CA
  pointmap(bin = floor(runif(n = length(myfips),1,5.99)), lat = bg.pts$lat[match(myfips, bg.pts$FIPS)], lon = bg.pts$lon[match(myfips, bg.pts$FIPS)], vartext = 'Example', areatext = 'all of CA',coloring = c('gray', 'light blue', 'dark blue', 'orange', 'red') , pch=1, cex=0.3)
}
\seealso{
\code{\link{countypointmap}}
}

