% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/proxistat.chunked.R
\name{proxistat.chunked}
\alias{proxistat.chunked}
\title{Call proxistat3 once per chunk & save output as file (breaks large input data into chunks)}
\usage{
proxistat.chunked(frompoints, topoints, fromchunksize, tochunksize,
  FUN = proxistat3, area, folder = getwd(), savechunks = TRUE,
  assemble = TRUE, ...)
}
\arguments{
\item{frompoints}{Require matrix or data.frame of lat/lon vauels that can be passed to get.distances3 function (colnames 'lat' and 'lon')}

\item{topoints}{Require matrix or data.frame of lat/lon vauels that can be passed to get.distances3 function (colnames 'lat' and 'lon')}

\item{fromchunksize}{Required, number specifying how many points to analyze at a time (per chunk).}

\item{tochunksize}{(not currently required - current default is to use all topoints at once) number specifying how many points to analyze at a time (per chunk).}

\item{FUN}{Optional function, \code{\link{proxistat3}} by default, and other values not implemented yet.}

\item{folder}{Optional path specifying where to save .RData files, default is getwd()}

\item{...}{Other parameters to pass to \code{\link{get.distances3}}, such as units}
}
\value{
Returns vector of character elements that are filenames for saved .RData output files in current working directory or specified folder.
  Each saved output is a vector of proximity scores if FUN=get.proxistat3, or matrix with extra columns depending on return. parameters above.
}
\description{
Call proxistat3 function in chunks, when list of frompoints is so long it taxes RAM (e.g. 11m blocks),
  saving each chunk as a separate .RData file in current working directory
}
\details{
filesizes if crosstab format (FASTEST & avoid needing rownums which take >twice as long & 1.25x sized file):
  80MB file/chunk if   1k blocks x 11k topoints/chunk:  y=get.distances.chunked(testpoints(11e6), testpoints(11000), 1e3, units='km',return.crosstab=TRUE)
  800MB file/chunk if 10k blocks x 11k topoints/chunk:  y=get.distances.chunked(testpoints(11e6), testpoints(11000), 1e4, units='km',return.crosstab=TRUE)
}

