% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speedsummary.R
\name{speedsummary}
\alias{speedsummary}
\title{Print summary of speed in buffers per hour, given start and end times of batch}
\usage{
speedsummary(start, end, n)
}
\arguments{
\item{start}{time when finished from Sys.time()}

\item{end}{time when finished from Sys.time()}

\item{n}{how many buffers were done?}
}
\value{
Prints to console and returns rate of buffers per hour
}
\description{
Print summary of speed in buffers per hour, given start and end times of batch
}
\examples{
# pts <- proxistat::testpoints_block2010(10)
# benchmark.start <- Sys.time()
# outlist <- proxistat::bufferapi(pts$lon, lat=pts$lat, radius = 5)
# benchmark.end <- Sys.time()
# proxistat::speedsummary(benchmark.start, benchmark.end, NROW(pts))
 # VISUALIZE variability in speed
 # hist(as.numeric(unlist(lapply(outlist, FUN =function(x) x$timeSeconds)), na.rm = T),100, 
 #   main = 'Histogram of seconds per buffer', xlab='Seconds elapsed for a buffer query of API')

}
